Player acceleration
Added player acceleration and deceleration.

Can be adjusted by changing the value of the acceleration variable (the
higher, the faster it takes to reach top speed, vice versa). For now I
set it to 35 (since moveSpeed is 7, it should take like, 0.2 seconds to
reach top speed, and 0.2 seconds to go from top speed to 0).